refactor(frontend): rename generate reports files
