+++ apps/frontend/utils/aggregate/generateIncidentReport.ts
@@ -1,6 +1,6 @@
 import { IncidentSpecific, KoboCommonKeys } from '@wfp-dmp/interfaces';
 
-import { aggregate } from '../aggregate';
+import { aggregate } from './aggregate';
 
 const firstKeys = [KoboCommonKeys.province, KoboCommonKeys.district];
